Revert 228002 "Enable cast_unittests"
Breaks compile on the `All` target on win_x64_rel. Note that this is only
built by default on manually submitted tryjobs. Not CQ or waterfall.
(see bug). For the errors see, e.g.,
http://build.chromium.org/p/tryserver.chromium/builders/win_x64_rel/builds/45000/

BUG=304877

> Enable cast_unittests
> And fixed associated errors
> 
> Review URL: https://codereview.chromium.org/26486005

TBR=pwestin@google.com

Review URL: https://codereview.chromium.org/27141002

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@228412 0039d316-1c4b-4281-b951-d872f2087c98

Former-commit-id: 166df683c229ed6c028afeb3e7bbe77f2565bda1

+++ media/cast/audio_sender/audio_encoder.cc
@@ -129,7 +129,7 @@ void AudioEncoder::EncodeAudioFrameThread(
     const FrameEncodedCallback& frame_encoded_callback,
     const base::Closure release_callback) {
   DCHECK(cast_environment_->CurrentlyOn(CastEnvironment::AUDIO_ENCODER));
-  size_t samples_per_10ms = audio_frame->frequency / 100;
+  int samples_per_10ms = audio_frame->frequency / 100;
   size_t number_of_10ms_blocks = audio_frame->samples.size() /
       (samples_per_10ms * audio_frame->channels);
   DCHECK(webrtc::AudioFrame::kMaxDataSizeSamples > samples_per_10ms)
